# perf(motion_utils): improve performance of findNearestIndex