Signed-off-by: Taiga Takano [taiga.takano@tier4.jp](mailto:taiga.takano@tier4.jp)

## Description
One of the major bottlenecks of this function is the execution of the atan function called within the calcYawDeviation function. Therefore, the order of execution of the conditional expressions is changed to reduce the likelihood that the calcYawDeviation function will be called.

## Before and After
This change speeds up the process as shown in the table below.

findNearestIndex callback latency
|  PERCENTILE  |  BEFORE (us)  |  AFTER (us)  |
| ---- | ---- | ---- |
|  99%  |  5.85  |  3.83  |
|  90%  |  3.81  |  3.36  |
|  50%  |  3.64  |  1.91  |